+++ src/MatrixDepot.jl
@@ -53,9 +53,9 @@ Access is like
 
 ###### commands:
     mdinfo, listdir, listgroups, matrixdepot, mdopen, listdata, mdlist,
-    metasymbols, loadsvd, setgroup!, deletegroup!.
+    metasymbols, setgroup!, deletegroup!.
 ###### selector patterns:
-    strings, string-patterns (using "*", "?", "/", "**"), regular expressions: for names
+    strings, string-patterns (using "*", "?", "[]", "/", "**"), regular expressions: for names
     builtin(42), user(3,5), sp(10:11,6,2833), mm(1), mm(:): to access by integer id or all
     sp(pattern), mm(pattern) to access corresponding (alternative) matrix for other collection
 ###### predicate patterns:
@@ -75,7 +75,6 @@ import Base: show
 
 export matrixdepot
 export listnames, listdir, listdata, listgroups, mdlist, mdinfo, metasymbols, mdopen
-export loadsvd
 export @addgroup, @rmgroup, @modifygroup # deprecated
 export setgroup!, deletegroup!
 
@@ -85,7 +84,7 @@ export isgeneral, issymmetric, isskew, ishermitian
 export iscomplex, isreal, isinteger, isboolean
 export isremote, islocal, isloaded, isunloaded, isbuiltin, isuser
 export issvdok, isposdef
-export @pred, keyword, hasdata
+export @pred, keyword, hasdata, charfun
 
 # The following functions are re-used as predicate functions / logical operators
 import Base: isreal, isinteger