+++ src/logical.jl
@@ -67,8 +67,8 @@ function aliasresolve(db::MatrixDatabase, a::Alias{T,<:AbstractVector}) where T
 end
 aliasresolve(db::MatrixDatabase, a::Alias{RemoteMatrixData{SSRemoteType},Colon}, ) = "*/*"
 aliasresolve(db::MatrixDatabase, a::Alias{RemoteMatrixData{MMRemoteType},Colon}) = "*/*/*"
-aliasresolve(db::MatrixDatabase, a::Alias{GeneratedMatrixData{:B},Colon}) = :builtin
-aliasresolve(db::MatrixDatabase, a::Alias{GeneratedMatrixData{:U},Colon}) = :user
+aliasresolve(::MatrixDatabase, ::Alias{GeneratedMatrixData{:B},Colon}) = :builtin
+aliasresolve(::MatrixDatabase, ::Alias{GeneratedMatrixData{:U},Colon}) = :user
 
 ###
 # Predefined predicates for MatrixData
@@ -111,6 +111,7 @@ example:
 """
 mm(p::P) where P<:Pattern = is_ivec(p) ? mm1(p) : Alternate{MMRemoteType,P}(p)
 
+is_ivec(::Integer) = true
 is_ivec(::AbstractVector{<:Integer}) = true
 is_ivec(p::AbstractVector) = all(is_ivec.(p))
 is_ivec(::typeof(:)) = true
@@ -161,7 +162,7 @@ islocal(data::GeneratedMatrixData) = true
 islocal(data::MatrixData) = false
 
 """
-    pred(f::Function, s::Symbol...)
+    pred_macro(f::Function, s::Symbol...)
 
 Return a predicate function, which assigns to a each `data::MatrixData`
 iff
@@ -169,7 +170,7 @@ iff
 *    all symbols `s` are property names of `data` and
 *    `f` applied to the tuple of values of those properties returns `true`
 """
-function pred(f::Function, s::Symbol...)
+function pred_macro(f::Function, s::Symbol...)
     data::MatrixData -> hasinfo(data) &&
     s ⊆ propertynames(data) &&
     f(getproperty.(Ref(data), s)...)
@@ -178,8 +179,11 @@ end
 """
     prednzdev(deviation)
 
-Test predicate - does number of stored (structural) non-zeros deviate from nnz
-by more than `deviation`. That would indicate a data error or high number of stored zeros. 
+Test predicate - does number of stored (structural) values deviate from nnz
+by more than `deviation`. That would indicate a data error or high number of stored zeros.
+
+The number of stored values is in `dnz`. It has to be approximately doubled for
+symmetric matrices to be comparable to `nnz``.
 """
 function prednzdev(dev::AbstractFloat=0.1)
     function f(data::RemoteMatrixData)
@@ -255,6 +259,7 @@ function extract_symbols(ex)
 end
 
 # construct a function definition from a list of symbols and expression
+# `make_func([:a, :b], expr)` returns `:( (a, b) -> expr )`
 function make_func(sli::AbstractVector{Symbol}, ex)
     res = :( () -> $ex )
     append!(res.args[1].args, sli)
@@ -270,7 +275,7 @@ const PROPS = (:name, :id, :metadata, fieldnames(MetaInfo)...)
 
 function make_pred(ex)
     syms = extract_symbols(ex) ∩ PROPS 
-    :(MatrixDepot.pred($(make_func(syms, ex)), $(QuoteNode.(syms)...)))
+    :(MatrixDepot.pred_macro($(make_func(syms, ex)), $(QuoteNode.(syms)...)))
 end
 
 """
@@ -287,3 +292,20 @@ macro pred(ex)
     esc(make_pred(ex))
 end
 
+"""
+    charfun(p::Pattern)
+
+Return the characteristic function corresponding to the set-defining pattern `p`.
+
+This function can be applied to objects `data::MatrixData` and strings, the canonical
+names of those objects.
+"""
+function charfun(p::Pattern)
+    f(d::MatrixData) = !isempty(list!(MATRIX_DB, [d.name], p)) # equivalent to: `d.name ∈ mdlist(p)`
+    f(s::AbstractString) = haskey(MATRIX_DB.data, s) && f(MATRIX_DB.data[s])
+    f
+end
+
+# make `:` an ordinary pattern (so it can be applied to MatrixData)
+import Base: (:)
+(:)(d::MatrixData) = true